Fix dry/wet & bypass with host in-place buffers
A LV2 plugin host may provide in-place buffers for processing.
ie (plug->input_l_p == plug->output_l_p). Processing
rewrites the input buffer.

However, rkr's x-fader bypass and dry/wet methods depend
on the input buffer to remain unchanged when processing.

To fix this, the input buffer is copied in case a host provides
inplace buffers.

(This does not yet add a special-cases for algorithms that
process in-place)